Remove coordinator markForAbortion calls
